move lod to 3d module
