+++ cocos/3d/index.ts
@@ -32,6 +32,7 @@ export * from './framework';
 
 export * from './lights';
 export * from './skinned-mesh-renderer';
+export * from './lod'
 export {
     utils,
 };
