+++ cocos/3d/lod/index.ts
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2018-2020 Xiamen Yaji Software Co., Ltd.
+ Copyright (c) 2022 Xiamen Yaji Software Co., Ltd.
 
  http://www.cocos.com
 
@@ -23,13 +23,4 @@
  THE SOFTWARE.
 */
 
-import './intersect';
-
-export { Camera } from './camera-component';
-export { ModelRenderer } from './model-renderer';
-export { Renderer } from './renderer';
-export { MissingScript } from './missing-script';
-export { PrefabLink } from './prefab-link';
-export { ReflectionProbe } from './reflection-probe-component';
-/** deprecated */
-export * from './deprecated';
+export { LOD, LODGroup } from './lodgroup-component';
